requirements.txt
fsspec==2023.12.2
gcsfs==2023.12.2.post1
google-cloud-storage==2.14.0
kaleido==0.2.1
numpy==1.26.3
pandas==1.5.3
pyrebase4==4.7.1
plotly==5.18.0
streamlit==1.29.0